# src/main/resources/static/openapi.yml
openapi: 3.1.0
info:
  title: Computer Architecture Course Administration API
  description: This API provides backend functionality for managing the Computer Architecture course system.
#  termsOfService: ?
#  contact:
#    name: ?
#    email: ?
#    url: ?
#  license:
#    name: ?
#    url: ?
  version: 1.0.0
#externalDocs:
#  description: ?
#  url: ?
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/v1/user:
    get:
      tags:
        - UserController
      summary: Retrieve a list of users
      description: >
        This endpoint allows you to fetch a list of users with optional sorting and filtering.
        
        - **Sorting** is available for all fields defined in `UserRequestDTO` (including `id`).
        
        - **Filtering** is available for `id` and `email` (can be used at the same time).
      operationId: getUsersByCriteria
      parameters:
        - $ref: '#/components/parameters/sort'
        - name: id
          in: query
          description: Filter users by ID.
          required: false
          schema:
            type: integer
            format: int32
        - name: email
          in: query
          description: Filter users by email address.
          required: false
          schema:
            type: string
            format: email
      responses:
        '200':
          description: OK - List of users returned.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponseDTO'
        '401':
          description: "Unauthorised - Unauthenticated: Authentication is required to access this resource."
        '403':
          description: "Forbidden - Access denied: You do not have permission to perform this action."
        '405':
          description: Method Not Allowed - Incorrect HTTP method used.
        default:
          description: Unexpected error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
    post:
      tags:
        - UserController
      summary: Create a new user/users
      description: Adds a new user/users to the system with the provided details.
      operationId: addUsers
      requestBody:
        description: Array of user objects that need to be created.
        content:
          application/json:
            schema:
              type: object
              properties:
                users:
                  type: array
                  items:
                    $ref: '#/components/schemas/UserRequestDTO'
        required: true
      responses:
        '201':
          description: Created - Users created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponseDTO'
        '400':
          description: Bad Request - Required request body is missing - Bad JSON format.
        '401':
          description: "Unauthorised - Unauthenticated: Authentication is required to access this resource."
        '403':
          description: "Forbidden - Access denied: You do not have permission to perform this action."
        '405':
          description: Method Not Allowed - Incorrect HTTP method used.
        '409':
          description: Conflict - The request could not be processed due to data integrity rules violation.
        '422':
          description: Unprocessable Entity - Validation Failed.
        default:
          description: Unexpected error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
    put:
      tags:
        - UserController
      summary: Update user by ID
      description: Updates information for an existing user identified by their ID.
      operationId: updateUserById
      parameters:
        - name: id
          in: path
          required: true
          description: The unique identifier of the user.
          schema:
            type: integer
            format: int32
      requestBody:
        description: User object with updated fields.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequestDTO'
        required: true
      responses:
        '200':
          description: OK - User updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponseDTO'
        '400':
          description: Bad Request - Required request body is missing - Bad JSON format.
        '401':
          description: "Unauthorised - Unauthenticated: Authentication is required to access this resource."
        '403':
          description: "Forbidden - Access denied: You do not have permission to perform this action."
        '404':
          description: Not Found - User not found.
        '405':
          description: Method Not Allowed - Incorrect HTTP method used.
        '409':
          description: Conflict - The request could not be processed due to data integrity rules violation.
        '422':
          description: Unprocessable Entity - Validation Failed.
        default:
          description: Unexpected error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
    delete:
      tags:
        - UserController
      summary: Delete user by ID
      description: Removes a user from the system identified by their ID.
      operationId: deleteUserById
      parameters:
        - name: id
          in: path
          required: true
          description: The unique identifier of the user.
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK - User deleted successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponseDTO'
        '401':
          description: "Unauthorised - Unauthenticated: Authentication is required to access this resource."
        '403':
          description: "Forbidden - Access denied: You do not have permission to perform this action."
        '404':
          description: Not Found - User not found.
        '405':
          description: Method Not Allowed - Incorrect HTTP method used.
        '409':
          description: Conflict - The request could not be processed due to data integrity rules violation.
        default:
          description: Unexpected error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
  /api/v1/user/password:
    put:
      tags:
        - UserController
      summary: Update user's password by ID
      description: >
        Updates the password of an existing user identified by their ID.  
        A new password is automatically generated by the system.
      operationId: updateUsersPasswordById
      parameters:
        - name: id
          in: path
          required: true
          description: The unique identifier of the user whose password is being updated.
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK - User's password updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponseDTO'
        '401':
          description: "Unauthorised - Unauthenticated: Authentication is required to access this resource."
        '403':
          description: "Forbidden - Access denied: You do not have permission to perform this action."
        '404':
          description: Not Found - User not found.
        '405':
          description: Method Not Allowed - Incorrect HTTP method used.
        '409':
          description: Conflict - The request could not be processed due to data integrity rules violation.
        default:
          description: Unexpected error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
  /api/v1/student:
    get:
      tags:
        - StudentController
      summary: Retrieve a list of students
      description: >
        This endpoint allows you to fetch a list of students with optional sorting and filtering.

        - **Sorting** is available for all fields defined in `StudentRequestDTO` (including `id`).

        - **Filtering** is available for `id`, `aisId` and `email` (can be used at the same time).
      operationId: getStudentsByCriteria
      parameters:
        - $ref: '#/components/parameters/sort'
        - name: id
          in: query
          description: Filter students by ID.
          required: false
          schema:
            type: integer
            format: int32
        - name: aisId
          in: query
          description: Filter students by AIS ID.
          required: false
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK - List of students returned.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StudentResponseDTO'
        '401':
          description: "Unauthorised - Unauthenticated: Authentication is required to access this resource."
        '403':
          description: "Forbidden - Access denied: You do not have permission to perform this action."
        '405':
          description: Method Not Allowed - Incorrect HTTP method used.
        default:
          description: Unexpected error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
    post:
      tags:
        - StudentController
      summary: Create and assign students to exercise
      description: Adds a new list of students and assigns them to the specified exercise.
      operationId: addStudents
      requestBody:
        description: Object containing exercise ID and list of students to be created and assigned.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentToExerciseRequestDTO'
        required: true
      responses:
        '201':
          description: Created - Students created successfully and assigned to exercise.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SuccessResponseDTO'
        '400':
          description: Bad Request - Required request body is missing - Bad JSON format.
        '401':
          description: "Unauthorised - Unauthenticated: Authentication is required to access this resource."
        '403':
          description: "Forbidden - Access denied: You do not have permission to perform this action."
        '405':
          description: Method Not Allowed - Incorrect HTTP method used.
        '409':
          description: Conflict - The request could not be processed due to data integrity rules violation.
        '422':
          description: Unprocessable Entity - Validation Failed.
        default:
          description: Unexpected error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
    put:
      tags:
        - StudentController
      summary: Update student by ID
      description: Updates information for an existing student identified by their ID.
      operationId: updateStudentById
      parameters:
        - name: id
          in: path
          required: true
          description: The unique identifier of the student.
          schema:
            type: integer
            format: int32
      requestBody:
        description: Student object with updated fields.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentRequestDTO'
        required: true
      responses:
        '200':
          description: OK - Student updated successfully.
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/SuccessResponseDTO'
        '400':
          description: Bad Request - Required request body is missing - Bad JSON format.
        '401':
          description: "Unauthorised - Unauthenticated: Authentication is required to access this resource."
        '403':
          description: "Forbidden - Access denied: You do not have permission to perform this action."
        '404':
          description: Not Found - Student not found.
        '405':
          description: Method Not Allowed - Incorrect HTTP method used.
        '409':
          description: Conflict - The request could not be processed due to data integrity rules violation.
        '422':
          description: Unprocessable Entity - Validation Failed.
        default:
          description: Unexpected error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
    delete:
      tags:
        - StudentController
      summary: Delete student by ID
      description: Removes a student from the system identified by their ID.
      operationId: deleteStudentById
      parameters:
        - name: id
          in: path
          required: true
          description: The unique identifier of the student.
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK - Student deleted successfully.
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/SuccessResponseDTO'
        '401':
          description: "Unauthorised - Unauthenticated: Authentication is required to access this resource."
        '403':
          description: "Forbidden - Access denied: You do not have permission to perform this action."
        '404':
          description: Not Found - Student not found.
        '405':
          description: Method Not Allowed - Incorrect HTTP method used.
        '409':
          description: Conflict - The request could not be processed due to data integrity rules violation.
        default:
          description: Unexpected error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
...
components:
  parameters:
    sort:
      name: sort
      in: query
      required: false
      description: >
        Sort the results by none or any field from the corresponding `Request/ResponseDTO`.
        Format: `field,(asc|desc)` >>> for example: `id,asc` or `id,desc`.
        If no order is specified, default is ascending.  
        Multiple sort parameters can be included, like:  
        `sort=field1,asc&sort=field2,desc`
      schema:
        type: string
  schemas:
...
    ErrorResponseDTO:
      type: object
      description: Represents a standard error response containing status code, general message, and optional list of detailed error messages.
      required:
        - status
        - message
      properties:
        status:
          type: integer
          format: int32
          description: HTTP status code representing the error.
        message:
          type: string
          description: A short description or summary of the error.
        errors:
          type: array
          items:
            type: string
          description: A list of detailed error messages.
    UserRequestDTO:
      type: object
      description: Request body used for creating new user or updating existing user.
      required:
        - fullName
        - email
        - roleEnum
      properties:
        fullName:
          type: string
          description: The user's full name (including degrees, first name, middle name, last name).
        email:
          type: string
          format: email
          description: The user's email address.
        roleEnum:
          type: string
          enum: [ ADMIN, TEACHER, HELPER ]
          description: The user's role within the system, determining access level and permissions.
    StudentRequestDTO:
      type: object
      description: Request body used for creating new student or updating existing student.
      required:
        - aisId
        - fullName
      properties:
        aisId:
          type: integer
          format: int32
          description: The student's AIS ID.
        fullName:
          type: string
          description: The student's full name (including degrees, first name, middle name, last name).
    StudentToExerciseRequestDTO:
      type: object
      description: Request body used for assigning multiple students to a specific exercise.
      required:
        - exerciseId
        - students
      properties:
        exerciseId:
          type: integer
          format: int32
          description: ID of the exercise to which the students will be added.
        students:
          type: array
          description: List of students to assign.
          items:
            $ref: '#/components/schemas/StudentRequestDTO'
...
    SuccessResponseDTO:
      type: object
      description: Represents a standard success response containing status code, general message, and optional list of data.
      required:
        - status
        - message
      properties:
        status:
          type: integer
          format: int32
          description: HTTP status code representing the success.
        message:
          type: string
          description: A short description or summary of the success.
        data:
          type: array
          items:
            type: object
          description: A list of updated or deleted data.
    UserResponseDTO:
      type: object
      description: Response body that returns details about a user.
      required:
        - id
        - fullName
        - email
        - roleEnum
      properties:
        id:
          type: integer
          format: int32
          description: The users's id.
        fullName:
          type: string
          description: The user's full name (including degrees, first name, middle name, last name).
        email:
          type: string
          format: email
          description: The user's email address.
        roleEnum:
          type: string
          enum: [ ADMIN, TEACHER, HELPER ]
          description: The user's role within the system, determining access level and permissions.
    StudentResponseDTO:
      type: object
      description: Response body that returns details about a student.
      required:
        - id
        - aisId
        - fullName
      properties:
        id:
          type: integer
          format: int32
          description: The student's id.
        aisId:
          type: integer
          format: int32
          description: The student's AIS ID.
        fullName:
          type: string
          description: The student's full name (including degrees, first name, middle name, last name).
...